chore: update deployment file names
