+++ scripts/chainOfTrustTransferRootManagerGM.ts
