+++ scripts/deployChainOfTrustBaseGM.ts
@@ -2,7 +2,7 @@ import { ethers, lacchain } from "hardhat";
 
 async function main() {
   const accounts = lacchain.getSigners();
-  const artifactName = "ChainOfTrustBase";
+  const artifactName = "ChainOfTrustBaseGM";
   const Artifact = await ethers.getContractFactory(artifactName, accounts[0]);
   console.log("Using Base Relay Address:", lacchain.baseRelayAddress);
   const revokeMode = 0; // only direct parent can revoke
