+++ scripts/deployChainOfTrustGM.ts
@@ -3,7 +3,7 @@ import { keccak256, toUtf8Bytes } from "ethers/lib/utils";
 
 async function main() {
   const accounts = lacchain.getSigners();
-  const artifactName = "ChainOfTrust";
+  const artifactName = "ChainOfTrustGM";
   const Artifact = await ethers.getContractFactory(artifactName, accounts[0]);
   console.log("Using Base Relay Address:", lacchain.baseRelayAddress);
   const revokeMode = 0; // only direct parent can revoke
