--- scripts/deployPublicDirectory.ts
@@ -1,21 +1,0 @@
-import { ethers, lacchain } from "hardhat";
-
-async function main() {
-  const accounts = lacchain.getSigners();
-  const artifactName = "PublicDirectory";
-  const Artifact = await ethers.getContractFactory(artifactName, accounts[0]);
-  console.log("Using Base Relay Address:", lacchain.baseRelayAddress);
-  const instance = await lacchain.deployContract(
-    Artifact,
-    lacchain.baseRelayAddress
-  );
-  console.log(
-    `${artifactName} instance successfully deployed at address: ` +
-      instance.address
-  );
-}
-
-main().catch((error) => {
-  console.error(error);
-  process.exitCode = 1;
-});
