+++ scripts/upgradeable/deployChainOfTrustBaseGMUpgradeable.ts
@@ -1,9 +1,8 @@
 import { ethers, lacchain, upgrades } from "hardhat";
-import { keccak256, toUtf8Bytes } from "ethers/lib/utils";
 
 async function main() {
   const accounts = lacchain.getSigners();
-  const artifactName = "ChainOfTrustUpgradeable";
+  const artifactName = "ChainOfTrustBaseGMUpgradeable";
   const deployer = accounts[0];
   const rootManager = accounts[0];
   console.log("signing with", deployer.address);
@@ -14,8 +13,6 @@ async function main() {
   const rootDid =
     "did:web:lacchain.id:3DArjNYv1q235YgLb2F7HEQmtmNncxu7qdXVnXvPx22e3UsX2RgNhHyhvZEw1Gb5C";
   const isRootMaintainer = false; // means only contract owner maintains "depth" and "revocation mode"
-  const didRegistryAddress = "0xd3dC5Bb0C17359D5ff130B186056eA8DC97a8D79";
-  const delegateType = keccak256(toUtf8Bytes("DefaultDelegateType"));
   await upgrades.deployProxy(
     Artifact,
     [
@@ -25,8 +22,6 @@ async function main() {
       rootManager.address, // root account manager
       revokeMode,
       isRootMaintainer,
-      didRegistryAddress,
-      delegateType,
     ],
     { kind: "uups" }
   );
