+++ scripts/upgradeable/deployPublicDirectoryGMUpgradeable.ts
@@ -2,11 +2,12 @@ import { ethers, lacchain, upgrades } from "hardhat";
 
 async function main() {
   const accounts = lacchain.getSigners();
-  const artifactName = "ExampleUpgradeable";
-  console.log("signing with", accounts[0].address);
-  const Artifact = await ethers.getContractFactory(artifactName, accounts[0]);
+  const artifactName = "PublicDirectoryGMUpgradeable";
+  const deployer = accounts[0];
+  console.log("signing with", deployer.address);
+  const Artifact = await ethers.getContractFactory(artifactName, deployer);
   console.log("Using Base Relay Address:", lacchain.baseRelayAddress);
-  await upgrades.deployProxy(Artifact, [4, lacchain.baseRelayAddress], {
+  await upgrades.deployProxy(Artifact, [lacchain.baseRelayAddress], {
     kind: "uups",
   });
   console.log("update addresses ..."); // TODO: address this issue
